third_party/WebKit/LayoutTests/fast/canvas/canvas-hit-regions-fill-rule-test.html

<!DOCTYPE html>
<html>
<head>
  <title>Canvas Hit Regions: fillRule test</title>
  <script src="../../resources/js-test.js"></script>
</head>
<body>
<canvas id="canvas" width="400" height="400">
</canvas>
<script src="./resources/canvas-hit-region-event.js"></script>
<script>

  var canvas = document.getElementById("canvas");
  var context = canvas.getContext("2d");

  context.rect(20, 20, 60, 60);
  context.rect(0, 0, 100, 100);
  context.addHitRegion({
    id : "nonzero"
    // default fillRule
  });

  debug("default fillRule(nonzero):");
  shouldBe("clickCanvas(1, 1)", "'nonzero'");
  shouldBe("clickCanvas(99, 98)", "'nonzero'");
  shouldBe("clickCanvas(21, 21)", "'nonzero'");
  shouldBe("clickCanvas(50, 50)", "'nonzero'");
  debug("");

  context.removeHitRegion("nonzero");

  context.addHitRegion({
    id : "evenodd",
    fillRule : "evenodd"
  });

  debug("fillRule = evenodd:");
  shouldBe("clickCanvas(1, 1)", "'evenodd'");
  shouldBe("clickCanvas(99, 98)", "'evenodd'");
  shouldBe("clickCanvas(21, 21)", "null");
  shouldBe("clickCanvas(50, 50)", "null");
  debug("");

  context.removeHitRegion("evenodd");

  context.addHitRegion({
    id : "nonzero",
    fillRule : "nonzero"
  });

  debug("fillRule = nonzero:");
  shouldBe("clickCanvas(1, 1)", "'nonzero'");
  shouldBe("clickCanvas(99, 98)", "'nonzero'");
  shouldBe("clickCanvas(21, 21)", "'nonzero'");
  shouldBe("clickCanvas(50, 50)", "'nonzero'");
  debug("");

  context.removeHitRegion("nonzero");

  shouldThrow('context.addHitRegion({id : "nonzero", fillRule : null });');

  context.addHitRegion({
    id : "nonzero",
    fillRule : undefined
  });

  debug("fillRule = undefined, falls back to default (nonzero):");
  shouldBe("clickCanvas(1, 1)", "'nonzero'");
  shouldBe("clickCanvas(99, 98)", "'nonzero'");
  shouldBe("clickCanvas(21, 21)", "'nonzero'");
  shouldBe("clickCanvas(50, 50)", "'nonzero'");
  debug("");

</script>
</body>
</html>
